set restart policy from task spec
we don't want to do this by default or for most containers because the
ip will probably change on restart and we'd have to add a bunch of new
code to detect that. instead just let orq start new containers like
before.

it's ok for nginx since we don't have update the nginx config to point
to a new ip.